Ajax ==> Update / DELETE / A faire : create

+++ src/Controller/TaskController.php
@@ -8,6 +8,8 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Routing\Annotation\Route;
 use App\Form\TaskType;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpFoundation\JsonResponse;
+
 
 class TaskController extends AbstractController
 {  /**
@@ -20,46 +22,23 @@ class TaskController extends AbstractController
         $entityManager = $this->getDoctrine()->getManager();
         // creates a task and gives it some dummy data for this example
         $task = new Task();
+        $duedate = new \DateTime('tomorrow');
+        $duedate -> format("Y-m-d");
+        $task->setDueDate($duedate);
         
-        $task->setDueDate(new \DateTime('tomorrow'));
-        /*   $tables[0]['first'] = "Alpha";
-        $tables[0]['seconde'] = "Beta";
-        $tables[0]['third'] = "Charlie";
-        
-        $tables[1]['first'] = "Alpha";
-        $tables[1]['seconde'] = "Beta";
-        $tables[1]['third'] = "Charlie"; */
         
         
         $form = $this->createForm(TaskType::class, $task);
         
         $form->handleRequest($request);
-        
-        /*  $contactFormData ="";
-        $taskName ="";
-        $dueDate="";
-        $task =[];
-        $contactFormDataObject =[];
-        $stockerObj= []; */
-        
+     
         if ($form->isSubmitted() && $form->isValid()) {
             
             
-            /*  $contactFormData =[];
-            foreach ( $form as $key => $value) {
-                $contactFormData[$key] = $value->getData();
-            } */
-            
+          
             
             $task = $form->getData(); 
-            /*   foreach ($contactFormDataObject as $key => $value) {
-                $stockerObj[$key] = $value;
-            } */
-            
-            /*  $taskName = $form["task"]->getData();
-            $dueDate = $form["dueDate"]->getData();
-            
-            $task =  $request->request->get('task'); */
+          
             $entityManager->persist($task);
             $entityManager->flush();
             $this->addFlash(
@@ -80,17 +59,14 @@ class TaskController extends AbstractController
         
         return $this->render('task/add.html.twig', [
             'form' => $form->createView(),
-            /* 'contactFormData' => $contactFormData,
-            'taskName' => $taskName,
-            'dueDate' => $dueDate,
-            'tasks' => $task,
-            'tables' => $tables,
-            'contactFormDataObject' => $contactFormDataObject,
-            'stockerObj' => $stockerObj, */
+          
             
             
             ]);
         }
+
+
+
         /**
         * @Route("/", name="task_show")
         */
@@ -115,58 +91,65 @@ class TaskController extends AbstractController
             
         }
         /**
-        * @Route("/task/update/{id}", name="task_update") 
+        * @Route("/task/update",options={"expose"=true}, name="task_update") 
         */
-        public function update(Request $request, $id)
+        public function update(Request $request)
         {
+            if ($request->isXmlHttpRequest()){
+                $data = json_decode(
+                    $request->getContent(),
+                    true
+                );
+            }
+            
+            $id = $data["id"];
+            $taskName = $data["taskname"];
+            $taskDueDate = $data["taskDueDate"]; 
             $entityManager = $this->getDoctrine()->getManager();
-            $task = $entityManager->getRepository(Task::class)->find($id);
+            $task = $entityManager->getRepository(Task::class)->find($id);    
+                
             
-            $form = $this->createForm(TaskType::class, $task);
+            $task->setTask($taskName);
+            $task->setdueDate(new \DateTime($taskDueDate));
+            $entityManager->flush();
+
+            return new JsonResponse($data);
+
             
-            $form->handleRequest($request);
-            
-            if ($form->isSubmitted() && $form->isValid()) {
-                
-                $task = $form->getData();
-                $taskName = $form["task"]->getData();
-                
-                
-                $dueDate = $form["dueDate"]->getData();
-                $task->setTask($taskName);
-                $task->setdueDate($dueDate);
-                $entityManager->flush();
-                return $this->redirectToRoute('task_show');
-                
-                
+        }
+        
+        
+        /**
+        * @Route("/task/delete",options={"expose"=true}, name="task_delete", methods={"POST"}) 
+        */
+        public function delete(Request $request)
+        {
+            if ($request->isXmlHttpRequest()){
+                $data = json_decode(
+                    $request->getContent(),
+                    true
+                );
             }
             
-            return $this->render('task/update.html.twig', [
-                'form' => $form->createView(), 
-                ]);
+            $id = $data;
+            if($id){
                 
-            }
-            /**
-            * @Route("/task/delete/{id}",options={"expose"=true}, name="task_delete") 
-            */
-            public function delete(Request $request, $id)
-            {
                 
-                if ($request->isXmlHttpRequest()){
-                $id=$request->request->get('id');
-                }
                 $entityManager = $this->getDoctrine()->getManager();
                 $task = $entityManager->getRepository(Task::class)->find($id);
                 
                 $entityManager->remove($task);
                 $entityManager->flush(); 
-                return $this->redirectToRoute('task_show');
+                
                 
                 
             }
-            
+            return new JsonResponse($id);
             
         }
         
         
-        
+    }
+    
+    
+
